Shared manifest-util tests dedicated crate
