+++ utils/manifest-utils/test-utils/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "ya-manifest-test-utils"
+version = "0.1.0"
+authors = ["Golem Factory <contact@golem.network>"]
+edition = "2018"
+homepage = "https://github.com/golemfactory/yagna"
+repository = "https://github.com/golemfactory/yagna"
+license = "LGPL-3.0"
+description = "Yagna manifest INTERNAL TEST utils"
+
+[dependencies]
+
+tar = "0.4"
+ya-manifest-utils = { path = ".." }
+
+ya-agreement-utils = "0.3"
