+++ utils/manifest-utils/test-utils/resources/test/certificates.tar
