+++ utils/manifest-utils/test-utils/resources/test/data.json.base64
