+++ utils/manifest-utils/test-utils/resources/test/data.json.base64.foo_req_sign.sha256.base64
