+++ utils/manifest-utils/tests/certificate_validation.rs
@@ -1,19 +1,19 @@
-pub mod common;
-
 #[macro_use]
 extern crate serial_test;
 
 use std::{fs, path::PathBuf};
 
-use common::*;
 use ya_manifest_utils::Keystore;
+use ya_manifest_test_utils::*;
+
+static TEST_RESOURCES: TestResources = TestResources { temp_dir: env!("CARGO_TARGET_TMPDIR") };
 
 #[test]
 #[serial]
 fn valid_certificate_test() {
     // Having
-    let (resource_cert_dir, test_cert_dir) = init_cert_dirs();
-    load_certificates(
+    let (resource_cert_dir, test_cert_dir) = TEST_RESOURCES.init_cert_dirs();
+    load_certificates_from_dir(
         &resource_cert_dir,
         &test_cert_dir,
         &["foo_ca-chain.cert.pem"],
@@ -32,8 +32,8 @@ fn valid_certificate_test() {
 #[serial]
 fn invalid_certificate_test() {
     // Having
-    let (resource_cert_dir, test_cert_dir) = init_cert_dirs();
-    load_certificates(&resource_cert_dir, &test_cert_dir, &[]);
+    let (resource_cert_dir, test_cert_dir) = TEST_RESOURCES.init_cert_dirs();
+    load_certificates_from_dir(&resource_cert_dir, &test_cert_dir, &[]);
 
     let request = prepare_request(resource_cert_dir);
 
@@ -58,7 +58,7 @@ struct SignedRequest {
 }
 
 fn prepare_request(resource_cert_dir: PathBuf) -> SignedRequest {
-    let resource_dir = common::test_resources_dir_path();
+    let resource_dir = TEST_RESOURCES.test_resources_dir_path();
 
     let mut cert = resource_cert_dir.clone();
     cert.push("foo_req.cert.pem");
